Combines essential fields into single string
The fields keyword, theme, title, and description are combined into a
single string for each dataset.

Cannot figure out how to avoid UnicodeEncodeError when writing to
a text file, the viewing of which can be helpful to understand how the
data is being manipulated.  The try except block
is the solution for now.

Sloppy but okay for now...